SQLite Database Functionality
Created a script that lets you fetch regulars. It will pull from a "regulars.txt" file and fill it into a database as the app starts. Keep in mind that there is no saving between sessions, all that matters is what is in the txt file. You can build on top of this to provide a bit of session management too.